Support agent-like proxy settings in the DCA
