+++ pkg/util/common.go
@@ -6,22 +6,15 @@
 package util
 
 import (
-	"crypto/tls"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"math/rand"
-	"net"
-	"net/http"
-	"net/url"
 	"os"
 	"path/filepath"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/util/log"
-
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
-	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/version"
 )
 
@@ -140,87 +133,4 @@ func GetJSONSerializableMap(m interface{}) interface{} {
 	}
 	return m
 
-}
-
-// GetProxyTransportFunc return a proxy function for a http.Transport that
-// would return the right proxy depending on the configuration.
-func GetProxyTransportFunc(p *config.Proxy) func(*http.Request) (*url.URL, error) {
-	return func(r *http.Request) (*url.URL, error) {
-		// check no_proxy list first
-		for _, host := range p.NoProxy {
-			if r.URL.Host == host {
-				log.Debugf("URL match no_proxy list item '%s': not using any proxy", host)
-				return nil, nil
-			}
-		}
-
-		// check proxy by scheme
-		confProxy := ""
-		if r.URL.Scheme == "http" {
-			confProxy = p.HTTP
-		} else if r.URL.Scheme == "https" {
-			confProxy = p.HTTPS
-		} else {
-			log.Warnf("Proxy configuration do not support scheme '%s'", r.URL.Scheme)
-		}
-
-		if confProxy != "" {
-			proxyURL, err := url.Parse(confProxy)
-			if err != nil {
-				err := fmt.Errorf("Could not parse the proxy URL for scheme %s from configuration: %s", r.URL.Scheme, err)
-				log.Error(err.Error())
-				return nil, err
-			}
-			userInfo := ""
-			if proxyURL.User != nil {
-				if _, isSet := proxyURL.User.Password(); isSet {
-					userInfo = "*****:*****@"
-				} else {
-					userInfo = "*****@"
-				}
-			}
-
-			log.Debugf("Using proxy %s://%s%s for URL '%s'", proxyURL.Scheme, userInfo, proxyURL.Host, SanitizeURL(r.URL.String()))
-			return proxyURL, nil
-		}
-
-		// no proxy set for this request
-		return nil, nil
-	}
-}
-
-// CreateHTTPTransport creates an *http.Transport for use in the agent
-func CreateHTTPTransport() *http.Transport {
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: config.Datadog.GetBool("skip_ssl_validation"),
-	}
-
-	if config.Datadog.GetBool("force_tls_12") {
-		tlsConfig.MinVersion = tls.VersionTLS12
-	}
-
-	// Most of the following timeouts are a copy of Golang http.DefaultTransport
-	// They are mostly used to act as safeguards in case we forget to add a general
-	// timeout to our http clients.
-	transport := &http.Transport{
-		TLSClientConfig: tlsConfig,
-		DialContext: (&net.Dialer{
-			Timeout: 30 * time.Second,
-			// Enables TCP keepalives to detect broken connections
-			KeepAlive: 30 * time.Second,
-			// Disable RFC 6555 Fast Fallback ("Happy Eyeballs")
-			FallbackDelay: -1 * time.Nanosecond,
-		}).DialContext,
-		MaxIdleConns:        100,
-		MaxIdleConnsPerHost: 5,
-		// This parameter is set to avoid connections sitting idle in the pool indefinitely
-		IdleConnTimeout:       90 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-	}
-
-	if proxies := config.GetProxies(); proxies != nil {
-		transport.Proxy = GetProxyTransportFunc(proxies)
-	}
-	return transport
 }
